Declare ansible_remote_user
The ansible_ssh_user is supposed to end up in the /etc/default/{{ env }}
file, however this variable is used internally to ansible and seems to
not be overriden for this template.

Rename it to ansible_remote_user instead. Fixes another misuse of
ansible_ssh_user to ansible_runner_user as well.