Add a role to install datadog agent
The agent is required on each host to be monitored. This initial bit of
code puts the agent in place and does the bare minimum configuration to
hook it into an account on Datadog. The api_key needs to be defined in
the localized secrets file.

This is the bare start enough to close the monitoring implementation
issue, and new issues can be created to address specific monitoring of
each service itself.

Fixes #5