Simple zookeeper role
If we are happy to use the ubuntu provided java and zookeeper then a
basic install isn't too hard. This is a super basic role that will need
customization to our requirements later. For now the templates are just
copy and pasted the default configs from the ubuntu packages.

Basically I wanted something I could play with for now and tune later.