Add datadog ansible callback
The callback plugin plus configuration go into ansible's virtualenv
path. The in-repo ansible.cfg is set to use that path, and if the path
doesn't exist, no error is thrown, so it's safe to run outside of our
bastion (however one can opt-in to having it send data from our personal
runs).

Fixes #32