Add support for getting secrets from a vault
This introduces the tooling to be able to use a vault encrypted file for
our secrets. A vault password file is defined in inventory, and will
have to be manually created on the runner host. The path to this file is
exposed as a ENV variable and is read by ansible(-playbook) at execution
time (rather than passed in as an explicit runtime option). For that
reason if a path is not defined in inventory variables, the ENV entry is
not created.

Related to #18