Add _meta back to dynamic inventory
In commit 0b5d9569e3c6e3ecf3775eaf3c68310c69ca1068 we removed the _meta
key from the inventory. However it's the meta key that provides the IP
addresses and other connection information that we use to actually
contact the servers. Without meta we end up getting errors when trying
to connect to UUID hosts.

Revert the change to this file.