Set a base ACL mask of r-- for secrets and vault pass
When diagnosing why the cideploy user couldn't read the vault pass I
discovered that the mask was empty meaning that there was no way to
provide any actual permissions to the object. We should always have the
secrets mask set to r-- so that the most that the greatest permissions
that can be provided to a user is read only.