Install user private ssh keys if they are provided in secrets
This adds a new 'ssh_keys' bucket to the secrets.yml which is used
to store private and public keys for users, both bastion users and
others.

bastion users who have an entry in the 'ssh_keys' dict will have their
private and public ssh keys installed into ~/.ssh. This intended to be
used to install the initial keys required by the cideploy user to do
the actual ansible deployments to remote nodes.

This is also used to store the zuul and nodepool keypairs. They are installed
into each node's user system user dirs, respectively.  Zuuls public key
is also used when creating the slave-authorized-keys file on nodepool,
which is baked into the slave images for ssh access.

This also fixes some syntax to match style we are using everywhere else.

Resolves: #33