Add Vagrantfile for local testing
Try to use vagrant to test a deploy of the bastion and then a deploy
of the full stack from the new bastion. This uses the example secrets
file, so nodepool will not be configured with a valid clouds.yaml and so
will not be functional. Also, this requires a valid cideploy keypair in
the example secrets file.