Add an allinone inventory set and config
This inventory file does everything on a single host. This is useful if
doing everything locally with -c local, or if one edits (the git
ignored) stub of an ssh config one can add the connection info for a
target host to deploy everything to. The allinone groupvars win over the
all groupvars if using the allinone inventory set.

The goal is to be able to actually run all the tasks with close to
production data to validate playbook contents beyond syntax check.