Add a travis job to run all the tasks locally
This should give us a better idea if our stuff will run to completion.
Later we can add more asserts that things did the stuff we wanted them
to do, or re-run the playbooks and look for any changes.

Because our code base requires Xenial (for systemd), we need to move to
Docker for testing the deployment. Travis only offers 14.04 as the
newest target. We have to run the container in a way that systemd
launches, so that our modules can do the right thing inside of them.