Add basic layout checks
This is a neutered version of the check-layout test infra uses to
validate that they haven't broken their zuul layout file. There are more
checks we can enable in this file that are currently disabled. As we
build out some job definitions and more repositories to test, we'll have
more in our layout.yaml and we will turn on more testing.