Fix nodepool pip install to only run with new commits
Currently, the pip task attempts a 'pip install --upgrade' on nodepool
every run. The pip module is apparently not smart enough to see that
there are no changes run-to-run, even if the code has not changed. This
is triggering the restart handler every time pip runs, so nodepool is
restarting every 15 minutes.

Now, pip will only be run if there are new commits to nodepool.