Fix Ansible's non-idempotent apt install for MySQL
Ansible 2.2 has a bug in the apt module that marks the task as changed,
if the cache is updated. This leads to it always being marked changed.

I've separated out the apt cache update from the package install, and
put in some explicit ok/fail code for the apt update. When Ansible 2.3
comes out, we can remove this silliness.