Add a gate for github projects
Some assumptions have been made on what a check vs a gate will look
like, this is a good enough starting point to be bikeshedded over later.

A gate check for github is triggered by adding the "approved" label.
This will start gate tests. If those tests pass it may try to merge, but
if github has branch protection on it won't merge. At the start of a
gate test it also removes the "approved" label making it easy to
re-trigger (by putting the label back on).

What we really want is to be able to trigger/require our own CI status
success, along with the approved label. However that will require
changes in zuul github code.