Os-clouds role should not set primary group to clouds.yaml group
Currently, when making sure that the owner of the clouds.yaml file can
read the file, it accidentally sets that owner's primary group to the
set clouds.yaml group. This has the side-effect of changing that user's
(cideploy) primary group when it is later created.

Instead, we should make sure that the owner is a member of the group by
adding that group to the additional groups of the owner.