Fix non-idempotent apt install for ansible-runner
Currently, every run of ansible-runner marks an apt module task as
changed, even when the only change is an apt cache update. This is
actually a bug in Ansible 2.2. I've separated out a separate task before
that attempts to update the cache separately, and fails if it can't, so
we should have the same behavior.