Add security group creation to create-hosts
Create some basic web security groups and add them to logs.
Unfortunately changing security groups can't be done via ansible so this
will only work when a server is first created. It also unfortunately
means that default is not applied automatically so we will have to
specify any internal interfaces as well.