Get project sources into nodepool node
With zuul 2.5 we have to tell zuul how to get the source code for a
project into the node. We will use zuul-cloner for this, which will take
advantage of a bunch of env vars to clone the right thing. But to use
zuul-cloner, we need it available in the node, and instead of installing
it as a build step, we should put it into the image. This is a stab at
doing it via elements.