Set zuul-cloner parameters
So the ZUUL_URL can only be used to fetch REFS and so doesn't have to be
bare, but it doesn't seem like you can clone from that url. We must
always specify a base URL. In this case we can base it from github.com
and find a more generic solution later.

I also change the path to zuul-cloner /usr/zuul-venv is something set up
by the openstack-infra modules and not where we install it to. For us
zuul is on the PATH so just use it.