Fix ansible-runner permissions issue
Since we have two ansible-runner crons with one running as root and the
other running as cideploy, we may end up in a situation where the
cideploy user does not have permission to update pip packages in the
ansible virtualenv. Run the pip command with sudo so that both crons are
able to update packages in the virtualenv.