Install os-loganalyze
os-loganalyze is openstack-infra's log formatter that adds some basic
formatting around the raw logs.

My takeaway from this is when we are serious we can do our own one of
these better than os-loganalyze. It's clunky to install, not pretty and
very basicly done.

For now however we can use this to display logs faster than writing our
own formatter.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>