Use service reload to recheck zuul-server config
There are two threads started by zuul-server, the main scheduler thread
and a gearman server thread. By using pkill to reload the zuul-server
configuration we are also sending SIGHUP to the gearman thread which
doesn't know how to handle the signal, dies, and generally causes a
bunch of problems.

We've already added an ExecReload= to the zuul-server systemd unit file
which we want to use instead to trigger a config refresh, however
because we use /bin/sh -c "..." in ExecStart the $MAINPID that we are
sending the SIGHUP to is the sh process which will kill the service.

To fix this we remove the /bin/sh in ExecStart= which means we can no
longer use ${PREFIX} because the path to the initial exec in a unit file
must be absolute and this causes an error.

To prevent a config reload being triggered when the server hasn't been
started we have an additional handler testing if the process is running
and only allowing a restart or reload notify to a running server. If
these fall through the services will be started after the flush_handlers
falls through.

After all this we should be able to have ansible correctly execute a
layout update without crashing zuul-server.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>