Add the start of a multinode hoist role.
To make multinode jobs work we seem to need a few things. First we need
a new node label in nodepool defining the layout.

After that we need to set up the new hosts to be able to ssh amongst
themselves by copying the generated ssh keys into place after nodepool
has started. This is done via nodepool scripts which are pre-baked into
an image.

Then we add the job definition and add it to a pipeline.

As it is it doesn't do anything. The next step will be to do a dynamic
ansible inventory that reads hosts out of /etc/nodepool/subnodes and
start to run the hoist role against it.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>