Make nodepool and zuul logs available via tailon
Create a monitor host, send nodepool and zuul logs there, and display them with
tailon.  Use Filebeat to ship the logs to a Logstash process on the new host.
The Logstash process then writes the logs to a file in a location where tailon
will find them. Using Filebeat and Logstash for this use-case makes it easier to
eventually expand to a full ELK cluster for system logs.

Since this is adding a requiremnets.yml, add logic for the file to our test jobs.

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>