Add bonnyci-run-pipeline job for executing in-repo test scripts
This adds a job template that can be re-used to run custom in-repo
job scripts for specified pipelines.

The job indicates to the script the pipeline in which the job is being run
via the 'BONNYCI_TEST_PIPELINE' environment variable.  It also exposes
a local directory via 'BONNYCI_TEST_LOG_DIR' which users may fill with
test output to be archived by the new 'bonnyci-logs' publisher.

The new job has been added to the sandbox repo for experimenting.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>