Make the hoist jobs fail on the first error
Run the jobs using `bash -ex` so that the first non-zero exit code fails the
entire job rather than running all commands to completion. The default `#!`
line is `#!/bin/bash -x`.

For the ubuntu-hoist job, this will prevent the validate playbook from running
if the deploy playbook fails, making the deploy failure easier to spot.

For the hoist-checks job, this will correctly treat any script failure
as a job failure, instead of only failing if the final script fails.

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>