Make sure the system hostname is in /etc/hosts
Scan /etc/hosts for a line starting with "127.0.0.1" and replace the line
with one listing both the short hostname (as reported by the system, up
to the first period) and "localhost".

The sudo utility will complain if it can't look up the system's hostname,
and adding a localhost alias in /etc/hosts is an easy way to avoid those
warnings.

Since docker can not edit /etc/hosts[1], add a tag for skipping tasks
docker can't perform. Also fix the docker commands to use the new 'monitoring'
tag that replaces the no-longer-existent 'ddapi' tag.

[1] https://github.com/docker/docker/issues/22281#issuecomment-214336587

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>