Add zuul_connections to all hosts
If we don't have the github connection in the deploy like in test
scenarios then the github connection isn't available in zuul and the
layout.yaml we add to all hosts isn't there. This means that the
layout.yaml fails validation and zuul doesn't start correctly.

Unfortunately zuul doesn't handle this failure well and instead of
failing completely gets stuck in a half loaded state where nothing
works.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>