Start of an ELK stack
So after copying most of the geerlingguy ansible roles I discovered they
were seriously out of date and that when updating it a lot of the config
wasn't the same in the new versions.

So the roles are basically the geerlingguy ones that have been
simplified down to what we actually use.

It works reasonably well for a first approximation, but there is a lot
to do around things like logstash filters and things to be actually
usable.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>