Set venv path fact from python-app
It is very common that after installing an application with python-app
that we want to use the venv directory that the app was installed to.
To do this with variables means having everyone aware of the /opt/venvs
prefix which is against the point of the role, but similarly just using
the python_app_venv_path variable would be overwritten if you are
installing multiple apps.

So set a fact in ansible that is unique to the application being
installed but known and able to be referenced by the depending role
later.

This is possibly the most flexible thing i've seen ansible do.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>