Fix multinode label assignment for BonnyCI jobs, clean noop tests
This explicitly defines node labels for the bonnyci-run-* jobs instead
of using job templating, which apparently does not work as I thought.

Also, this jobs the test stubs that we needed to put in place to get
the first commit through the gate, and it drops unused job config +
some travis tests that are covered by Bonny now.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>