Add automation to create dev users/projects
This is a playbook that's meant to be executed by a developer, on a
system that has credentials suitable to create projects and users
defined in a cloud.yaml file. The name of the cloud can be overridden by
setting a runtime variable of "prod_cloud", otherwise it'll default to
"opentech-sjc".

Each user will get a project that matches their username, and get
_member_ role within. Each user will also be added to the prod group,
which defaults to "bonnyci". This group has _member_ role on a
specific production project, allowing each member to see resources
managed by the production account. Select members can also be added to
the production admin account, "bonnyci-admin", allowing such members to
be able to do cloud_admin level things should they need to.

This commit also vendors in a newer os_user.py from upstream Ansible, so
that the password will not be updated if the user already exists. This
allows us to set a default password on initial creation, which will
quickly be changed.

Closes BonnyCI/projman#131

Signed-off-by: Jesse Keating <jkeating@j2solutions.net>