Move apache status page to a different port
The apache status module helpfully installs itself by default onto all
vhosts, but restricts itself to localhost. This is questionable
behaviour anyway but it does not play nicely with our log rendering on
the logs server and so dd-apache is failing on that machine.

Construct a new vhost on a random port that only listens on localhost
and serves the status url. Then point the datadog check to this port
instead.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>