Port over create-hosts to new cloud-bootstrap + cloud-ready roles
This ports over the code we used to create new hosts from the
create-hosts role into the new, more modular cloud-bootstrap and
cloud-ready roles for provisioning prod and dev environments.

2 SSH keys get created:

 * A default key (specified in cloud-bootstrap defaults) that is
   used to boot and initially access the bastion
 * A cideploy key (specified via secrets) that is used to boot
   other nodes that will be under management of automation on
   the bastion's.

The cloud-bootstrap role will boot instances with specific secgroups,
and associate floating IPs either from a network or using a specific IP.

After the cloud-bootstrap role has run, prod environments are expected
to update DNS for the domain specified in dns_subdomain, so that the
FQDNs of each instance (as specified by instance name) point to the
private ipv4 address of each.  For developer environments, this is not
necessary.

After DNS has been updated, the cloud-ready role may be run against
a bastion to complete final tasks necessary prior to deploying the
bastion itself.  This includes ssh keyscans of private addresses
and hostnames.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>