Create scp-log-receiver
There are a few places between the zuul and logging servers where we
rely on coordinating a place where the launcher can upload logs to. We
need not only to have a user that can perform this operation, but a
directory with appropriate permissions.

Refactor this and make it so all environments share common site
information, but with their own scp log host.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>