Defined zuul_sites per environment
zuul_sites becomes the sites that the jobs can specify to upload logs
to. At the moment because this is defined in the role even test
environments are trying to upload to the production zuul logs server.

Set the role default to not produce any zuul sites, copy the existing
information to production and set up some sensible defaults for the
other environments.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>