Move old letsencrypt cert when requesting a new one
The letsencrypt module will check the validity of the current cert before
requesting a new one, without checking that the cert and key match. If a
new key and cert request are generated while the old cert is still valid,
this leaves us with a key/cert mismatch and apache will fail to start.

If we generate a new key and/or cert request, move the old cert to a backup
location to force the letsencrypt module to generate a new cert.

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>