ansible-runner: Use the correct pip
This currently calls the system pip to install requirements,
instead of installing them into the ansible venv.  I think
we somehow have not hit this because the venv in production
already had dependencies installed into it at one point in
the past.

It also moves it later in execution, to ensure we run the installation
after we've updated the repo and pulled in any new requirements.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>