Add tools/vagrant-deploy.sh
Previously, vagrant deployments were not happening the same way that we
deploy in production. Each time a new host was brought up, it would have
ansible run against it and it alone. Since, our playbooks are written
with the assumption that they will hit all hosts in a certain order,
this method left room for improvement.

Now, vagrant deployments follow the same process as our production
deployments (not to mention our CI deploys, and the proposed docker
deploys). It first brings up all of the hosts, then the bastion server
runs the entire ansible playbook for a deployment, followed by
validation playbook.

Related-Issue: BonnyCI/projman#172
Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>