Fix quartermaster service's working directory
Currently, the working directory for quartermaster is whatever
systemd says it is. It would be nice for plugins to be able to use
relative paths, so they don't need knowledge of where they are
installed. This changes the systemd service file to use the
quartermaster user's homedir as the working dir for the bot.

Signed-off-by: Bradon Kanyid <bradon@kanyid.org>