Slice up provisioning to support multiple zuul environments per bastion
This creates a common network and puts services that are shared among
environments on it, including the bastion.

It then moves per-environment hosts in to an 'environments' list that
will allow easily expanding our infra to include many zuul environments.

A reasonable set of defaults are used so that individual users can spin up
similar environments in their tenant space.  Those may be overriden to supply
site-specific values for production.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>