Manage multiple nodepool projects, set quotas for prod accounts
This updates account management to allow creation of multiple nodepool
projects to be used across multiple zuul/nodepool environments.

It also adds the ability to set quotas for production and nodepool
accounts, defaulting to opening them up to unlimited usage.  It
vendors a new, unreleased os_quota.py module in library/ which includes
some fixes to a bug we hit when trying to use it on our cloud.

This also fixes a typo in a defaults file.

Related-Issue: BonnyCI/projman#139
Closes-Issue: BonnyCI/projman#194

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>