Allow multiple cron jobs per ansible-runner task
Instead of requiring multiple ansible-runner tasks at the top-level,
allow a list of jobs to passed in as a variable.  We currently do this at
the top-level play but that makes it difficult to run different sets of
jobs for different environments.  This will allow us to run many runner
jobs in production with different inventories.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>