Fix missing user + ansible_remote_user distinction
I, Adam Gandelman, broke ansible-runner tasks in a previous commit
by not distinguishing between the local user running the task and
the remote user that is written to the tasks' /etc/default/ config
file.

This updates it to write out based on the optional ansible_remote_user
variable.  It also updates the task to use the correct inventory for
the bastion run.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>