Drop Zuul V3 multinode label, decrease min-ready, update provider
Our cloud is running out of resources. We don't initially need any resources
for our V3 testing.  This drops the v3 min-ready to 1, drops the multinode
label entirely and uses a smaller instance size for the slaves. It decreases
the production min-ready.  We can up this as we onboard projects and our
usage and queue times increase.

It also updates the node provider name for the V3 environment, which gets set
in slave instances' metadata and is used by nodepool to filter instances.  This
should be unique.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>