Nodepool-v3 network should create nodes with open access
Nodepool will try to ssh to the newly created nodes to verify that they
are up. To do this they need to be in a security group that allows it.
The way this is handled in v2 is that the default sg just allows access
to all ports from the control plane network. This is probably
unnecessary and should be just SSH, but replicate it for now.

This has not actually been tested because I don't really have the
ability, but it reflects a change I had to make on the cloud.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>