Add a default deny apache vhost to elk
A previous commit that was meant to setup kibana reverse proxying
mistakenly opened the host up as a public forward proxy, and was
picked up as a backend for a public proxying service. Forward
proxying was disabled, but we're still getting tons of incoming
proxy requests.  This sets up a default vhost to deny these and
only accept requests explicitly bound for configured vhosts.  This
will hopefully result in the address being pulled from the proxy
service's database.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>