Add new inventory for travis bastion playbook
Move the allinone bastion host entry to its own inventory file
so that the various CI group_vars do not apply to the bastion
host. This is needed so that the apache fail2ban jails are not
enabled on the bastion host since we are not installing apache
on the bastion -- and enabling those jails causes fail2ban to
fail to start.

Also, make sure /var/log/auth.log exists since that is used by
the default sshd jail.

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>