Fix handling of force option in ansible-runner
Previously, the --force option was being passed to ansible after being
used to force the ansible-runner script because ansible uses the
deprecated optparse, which allows for abbreviated options. As a result
`--force` was being interpreted by ansible-playbook as
`--force-handlers`.

Now, ansible-runner uses the more obvious `--force-runner` (per
@gandelman-a's request), which switches on `force_runner`. All other
arguments are moved to the `ansible_options` string, which is in turn
passed to `ansible-playbook`.

With these changes made, run-one breaks the script. However, our
deployments are currently unlikely to be impacted by us not using
run-one. As such, it has been removed.

Related-Issue: BonnyCI/projman#211
Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>