Add a second run to test-hoist-ansible.sh
Previously, we were only running install-ci.yml once during CI. Doing so
did not allow for us to check the number of items which are changed on
subsequent runs.

Now we run install-ci.yml a second time and verify that the number of
changed items is what we expect.

Related-Issue: BonnyCI/projman#49
Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>