Reduce min-ready to 1 for both labels
Since we have a small cloud quota, it's easy for the ubuntu-hoist label
to take up all of our resources since it uses multiple nodes, leaving us
with no ubuntu-xenial nodes ready. Drop both min-ready counts to 1 to make
it more likely that we will ready nodes of both labels.

This means we may spend more time waiting for nodes to build.

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>