Add script to perform dib image validation
We should check for some necessities during the image build process and
fail if any of them are not found or non-functional. This will add a
script that runs late in the build process and checks:

  * python 2.7.x installation
  * virtualenv installation
  * zuul-cloner check

As it stands right now, the checks are fairly rudimentary and just
ensure that the commands run without error.

Implements: BonnyCI/projman#215

Signed-off-by: Cullen Taylor <cullentaylor@outlook.com>