Add a way to disable the ansible-runners from ansible
Add a disable flag that disables the ansible-runner from running on
cron. This is useful for testing where I don't want all the runners to
actually start.

I haven't added the reverse that removes the flag from git as I feel you
probably want to control this manually so the automation doesn't remove
your manual hold.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>