Add pruner script to ARA role
ARA currently is saving entries to a sqlite db on the bastion. We run
ansible *a lot* so this database is getting quite large. This script
uses ARA's cli tool to determine ara playbooks older than a certain
date, and deletes them from the database.

This is a stop-gap solution until I move the ARA database to mysql on
the logs host.

Signed-off-by: Bradon Kanyid <bradon@kanyid.org>