Add logging-proxy deployment to logs server
At the moment we deploy the logging proxy onto the log server on a high
port. We don't need to expose this on a standard port for now because
people will only follow links. In future the client side of this can be
integrated into the bonnyci.org site and the websock be on whatever port
we like. For now splitting that up is not worth it.

I've set up the proxy so that it redirects through apache so that when
we put SSL on the site this will also be affected.

There's some logging magic in here to get from journald -> syslog ->
file so we can later put it into logstash.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>