Use ansible_hostname as datadog hostname
I think it's still correct in the dd-agent not to automatically set
hostname because for a lot of cases the system derived hostname is
probably correct. For us who created a lot of short hostnames in
openstack that come through as the hostname in datadog this means that
the datadog ansible hosts don't line up with the actual hosts.

Always default to the ansible set hostname in datadog.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>