Standardize hostnames in inventory
When ansible hits the datadog callback it uses the inventory hostname as
the machine it is registering. This means we have each host in datadog
defined twice, once for the dd-agent it is running and once for the
ansible that is configuring it. I'd bet we're paying twice as well.

Work on a standardized host naming system that we can use across
datadog, filebeat and ansible to name all the hosts the same thing.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>