Revert 'ansible-runner: Drop flag files for failing environments'
This is no longer required, as the datadog checks were fixed that trigger on
data we send via the datadog callback.  I should have abandoned this change
when that happened, my bad.

Also, there's a bug in the change that this reverts where runners that
are not running as root cannot create the flag file.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>