Add an ansible-runner task to run bonnyci-test-suite
Until we can run this as a periodic job in Zuul and get its logs
published into logstash, run it as an ansible task from the bastion.

This essentially just sets up the task, a user and passes through some
secrets.  The test suite itself contains a playbook, which the ansible-runner
task calls, to convert the secrets into a test config and run the test suite.

The datadog monitor should be able to monitor for the runner task and
report on its failure.

Closes-Issue: BonnyCI/projman#178

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>