Export ARA_DATABASE in ansible-runner for ansible
I thought that having ARA_DATABASE available in the ansible-runner
script would pass those env vars to the ansible process that is started
from that script. Apparently not.

This exports the var so it is in ansible's environment. It unexports
itself at the end of the script, so it doesn't linger in the caller's
shell.

Signed-off-by: Bradon Kanyid <bradon@kanyid.org>