Fixup vagrant port forwards
Match the 808X port to the .10X ip address. This was already the case
for most 808X ports, but both logs and backups were trying to use port
8085. Since backups uses ip address .105, use the port on that host. The
logs host uses ip address .102 so use port 8082 on that host.

Also add a direct port forward for kibana to bypass the apache reverse
proxy in front of it.

Signed-off-by: K Jonathan Harker <Jonathan.Harker@ibm.com>