Update secrets file handling in vagrant
Previously, provisioning the bastion VM in vagrant would put what
whatever was in `secrets.yml` in place on the bastion, causing many our
local tests to be rendered inaccurate.

Now, if there is a file at `hoist/secrets.yml`, that is used, but if
there is no file there, we use `hoist/secrets.yml.example`. In effect,
we will no longer have to manually remove `hoist/secrets.yml` everytime
there is a change to the example file.

NOTE: Anyone who has an old copy of secrets.yml from before this change
will still be using that one. If you do not have a reason for using a
customized secrets file, you should remove the file at
`hoist/secrets.yml`.

Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>